[ADD/CHORE] 색추가 & 닉네임 디자인 수정 Extension에서 오렌지 색 추가와 닉네임 PR 중 디자인 수정
