+++ AppleStop/AppleStop/Global/Extension/Color+Extension.swift
@@ -19,6 +19,12 @@ extension Color {
     static var mainGreen: Color {
         return Color(hex: "#51AA6F")
     }
+    static var iconGrey: Color{
+        return Color(hex: "#8A8A8A")
+    }
+    static var charOrange: Color{
+        return Color(hex: "#f26725")
+    }
 }
 
 extension Color {
